requirements.txt
django==1.9.6
Pillow
django-registration
django-linguist
django-rosetta
#sudo apt-get install graphviz libgraphviz-dev graphviz-dev python-pygraphviz
django-extensions
